Move Jasper() class definition into application module
